requirements.txt
Collectfast==0.2.1
Django==1.7.1
Fabric==1.10.1
Markdown==2.5.2
Pillow==2.6.1
PyYAML==3.11
Pygments==2.0.1
SQLAlchemy==0.9.8
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
bcrypt==1.1.0
billiard==3.3.0.19
boto==2.34.0
celery==3.1.17
cffi==0.8.6
closure==20140110
coverage==3.7.1
django-activeurl==0.1.8
django-appconf==0.6
django-bower==5.0.1
django-cached-authentication-middleware==0.2.0
django-celery==3.1.16
django-celery-email==1.0.4
django-classy-tags==0.5.2
django-clear-cache==0.3
django-compressor==1.4
django-configurations==0.8
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-debug-toolbar-template-timings==0.6.4
django-extensions==1.4.9
django-guardian==1.2.4
django-heroku-memcacheify==0.8
django-mailviews==0.6.3
django-memcached-pool==0.5
django-mobile==0.4.0
django-multiurl==1.0.1
django-phonenumber-field==0.6
django-pickling==0.1
django-pylibmc==0.5.0
django-pylibmc-sasl==0.2.4
django-redis-sessions-fork==0.6.2
django-rest-swagger==0.2.6
django-reversion==1.8.5
django-secure==1.0.1
django-storages==1.1.8
django-transaction-hooks==0.2
djangorestframework==3.0.2
docutils==0.12
drf-extensions==0.2.6
drf-ujson==1.1
ecdsa==0.11
flake8==2.2.5
geoip2==2.1.0
gnureadline==6.3.3
gunicorn==19.1.1
hiredis==0.1.5
honcho==0.5.0
httpie==0.8.0
ipaddr==2.1.11
ipdb==0.8
ipython==2.3.1
kombu==3.0.24
librabbitmq==1.6.1
lxml==3.4.1
maxminddb==1.1.1
mccabe==0.3
newrelic==2.40.0.34
nltk==3.0.0
nose==1.3.4
paramiko==1.15.2
pep8==1.5.7
phonenumbers==7.0.1
pip-tools==0.3.5
psycopg2==2.5.4
py==1.4.26
pycparser==2.10
pycrypto==2.6.1
pyflakes==0.8.1
pylibmc==1.4.1
python-dateutil==2.3
pytz==2014.10
raven==5.1.1
redis==2.10.3
regex==2014.12.15
requests==2.5.0
six==1.8.0
sqlparse==0.1.14
tox==1.8.1
uWSGI==2.0.8
ujson==1.33
umemcache==1.6.3
virtualenv==1.11.6
wsgiref==0.1.2
yuicompressor==2.4.8
